♻️ REFACTOR: cardlist 안에 들어있는 여러 컴포넌트들 개별화 시키기
